Rename ClassificationInput to ClassificationInputMixin
