+++ flash/core/integrations/icevision/data.py
@@ -16,7 +16,7 @@ from typing import Any, Callable, Dict, List, Optional, Type, Union
 
 import numpy as np
 
-from flash.core.data.io.classification_input import ClassificationState
+from flash.core.data.io.classification import ClassificationState
 from flash.core.data.io.input import DataKeys, Input
 from flash.core.data.utilities.paths import list_valid_files
 from flash.core.integrations.icevision.transforms import from_icevision_record
