+++ flash/image/detection/output.py
@@ -15,7 +15,7 @@ from typing import Any, Dict, List, Optional, Union
 
 from pytorch_lightning.utilities import rank_zero_warn
 
-from flash.core.data.io.classification_input import ClassificationState
+from flash.core.data.io.classification import ClassificationState
 from flash.core.data.io.input import DataKeys
 from flash.core.data.io.output import Output
 from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, lazy_import, requires
