+++ flash/template/classification/data.py
@@ -20,7 +20,7 @@ from flash.core.data.base_viz import BaseVisualization
 from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
-from flash.core.data.io.classification_input import ClassificationInput
+from flash.core.data.io.classification import ClassificationInputMixin
 from flash.core.data.io.input import DataKeys, Input
 from flash.core.data.io.input_transform import InputTransform
 from flash.core.data.utilities.samples import to_samples
@@ -34,7 +34,7 @@ else:
     Bunch = object
 
 
-class TemplateNumpyClassificationInput(ClassificationInput):
+class TemplateNumpyClassificationInput(Input, ClassificationInputMixin):
     """An example data source that records ``num_features`` on the dataset."""
 
     def load_data(
